+++ lib/global_uid/error_tracker.rb
@@ -1,0 +1,7 @@
+module GlobalUid
+  class ErrorTracker
+    def notify(exception)
+      ActiveRecord::Base.logger.error("GlobalUID error: #{exception.class} #{exception.message}")
+    end
+  end
+end