+++ Changelog.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - [Breaking change] ID Validation, ensure the ID coming back has been incremented using the configured `auto_increment_increment`. (https://github.com/zendesk/global_uid/pull/63)
 - [Breaking change] Move configuration to a class (https://github.com/zendesk/global_uid/pull/70)
 
+## Changed
+- [Breaking change] `notifier` configuration option now only takes an exception, rather than exception and message
+
 ### Removed
 - Removed the `dry_run` option (https://github.com/zendesk/global_uid/pull/64)
 - Removed `GlobalUid::ServerVariables` module (https://github.com/zendesk/global_uid/pull/66)